refactor(dsmiddleware/splitop): rename package splitcall to splitop
